Change rating type.

+++ app/src/main/java/com/example/richo_han/richonews/MainActivity.java
@@ -47,7 +47,7 @@ public class MainActivity extends AppCompatActivity {
             if(resultCode == RESULT_OK) {
                 ListView listView = (ListView) findViewById(R.id.news_list);
                 int index = data.getIntExtra(NewsContentActivity.EXTRA_NEWS_INDEX, 0);
-                int rating = data.getIntExtra(NewsContentActivity.EXTRA_NEWS_RATING, 0);
+                float rating = data.getFloatExtra(NewsContentActivity.EXTRA_NEWS_RATING, 0);
                 View view = listView.getChildAt(index);
                 ((RatingBar) view.findViewById(R.id.ratingBar)).setRating(rating);
             }
